+++ tests/test_update.py
@@ -1,12 +1,12 @@
-# pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring
+# -*- coding: utf8 -*-
+# pylint: disable=missing-function-docstring
 
-from __future__ import unicode_literals
+"""Tests for metadata update commands."""
 
 import json
 
 from binstar_client import errors
-from binstar_client.scripts.cli import main
-from tests.fixture import CLITestCase
+from tests.fixture import CLITestCase, main
 from tests.urlmock import urlpatch
 from tests.utils.utils import data_dir
 
@@ -30,12 +30,13 @@ release_test_data = {
 
 
 class TestUpdate(CLITestCase):
+    """Tests for metadata update commands."""
 
     @urlpatch
     def test_update_package_from_json(self, urls):
         urls.register(method='HEAD', path='/', status=200)
         update = urls.register(method='PATCH', path='/package/owner/package_name', content=json_test_data, status=200)
-        main(['update', 'owner/package_name', data_dir('metadata.json')], False)
+        main(['update', 'owner/package_name', data_dir('metadata.json')])
 
         urls.assertAllCalled()
         req = json.loads(update.req.body)
@@ -46,7 +47,7 @@ class TestUpdate(CLITestCase):
         urls.register(method='HEAD', path='/', status=200)
         update = urls.register(method='PATCH', path='/package/owner/package_name',
                                content=package_test_data, status=200)
-        main(['update', 'owner/package_name', data_dir('test_package34-0.3.1.tar.gz')], False)
+        main(['update', 'owner/package_name', data_dir('test_package34-0.3.1.tar.gz')])
 
         urls.assertAllCalled()
         req = json.loads(update.req.body)
@@ -57,7 +58,7 @@ class TestUpdate(CLITestCase):
         urls.register(method='HEAD', path='/', status=200)
         update = urls.register(method='PATCH', path='/release/owner/package_name/1.0.0',
                                content=json_test_data, status=200)
-        main(['update', 'owner/package_name/1.0.0', data_dir('metadata.json'), '--release'], False)
+        main(['update', 'owner/package_name/1.0.0', data_dir('metadata.json'), '--release'])
 
         urls.assertAllCalled()
         req = json.loads(update.req.body)
@@ -68,7 +69,7 @@ class TestUpdate(CLITestCase):
         urls.register(method='HEAD', path='/', status=200)
         update = urls.register(method='PATCH', path='/release/owner/package_name/1.0.0',
                                content=release_test_data, status=200)
-        main(['update', 'owner/package_name/1.0.0', data_dir('test_package34-0.3.1.tar.gz'), '--release'], False)
+        main(['update', 'owner/package_name/1.0.0', data_dir('test_package34-0.3.1.tar.gz'), '--release'])
 
         urls.assertAllCalled()
         req = json.loads(update.req.body)
@@ -80,7 +81,7 @@ class TestUpdate(CLITestCase):
         urls.register(method='PATCH', path='/package/owner/package_name', content=package_test_data, status=404)
 
         with self.assertRaises(SystemExit):
-            main(['update', 'owner/package_name', data_dir('not_existing.tar.gz')], False)
+            main(['update', 'owner/package_name', data_dir('not_existing.tar.gz')])
 
     @urlpatch
     def test_update_package_not_found(self, urls):
@@ -88,14 +89,14 @@ class TestUpdate(CLITestCase):
         urls.register(method='PATCH', path='/package/owner/package_name', content=package_test_data, status=404)
 
         with self.assertRaises(errors.NotFound):
-            main(['update', 'owner/package_name', data_dir('test_package34-0.3.1.tar.gz')], False)
+            main(['update', 'owner/package_name', data_dir('test_package34-0.3.1.tar.gz')])
 
     @urlpatch
     def test_update_release_missing_version(self, urls):
         urls.register(method='HEAD', path='/', status=200)
         urls.register(method='PATCH', path='/release/owner/package_name/1.0.0', content=package_test_data, status=200)
         with self.assertRaises(errors.UserError):
-            main(['update', 'owner/package_name', data_dir('test_package34-0.3.1.tar.gz'), '--release'], False)
+            main(['update', 'owner/package_name', data_dir('test_package34-0.3.1.tar.gz'), '--release'])
 
     @urlpatch
     def test_update_release_not_found(self, urls):
@@ -103,4 +104,4 @@ class TestUpdate(CLITestCase):
         urls.register(method='PATCH', path='/release/owner/package_name/1.0.0', content=package_test_data, status=404)
 
         with self.assertRaises(errors.NotFound):
-            main(['update', 'owner/package_name/1.0.0', data_dir('test_package34-0.3.1.tar.gz'), '--release'], False)
+            main(['update', 'owner/package_name/1.0.0', data_dir('test_package34-0.3.1.tar.gz'), '--release'])