+++ tests/test_whoami.py
@@ -1,31 +1,26 @@
-# pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring
+# -*- coding: utf8 -*-
+# pylint: disable=missing-function-docstring
 
-"""
-Created on Feb 18, 2014.
-
-@author: sean
-"""
-
-from __future__ import unicode_literals
+"""Tests for whoami command."""
 
 import json
 import os
-import unittest
-from unittest import mock
+import unittest.mock
 
 import requests.utils
 
-from binstar_client.scripts.cli import main
-from tests.fixture import CLITestCase
+from tests.fixture import CLITestCase, main
 from tests.urlmock import urlpatch
 from tests.utils.utils import data_dir
 
 
 class Test(CLITestCase):
+    """Tests for whoami command."""
+
     @urlpatch
     def test_whoami_anon(self, urls):
         user = urls.register(method='GET', path='/user', status=401)
-        main(['--show-traceback', 'whoami'], False)
+        main(['--show-traceback', 'whoami'])
         self.assertIn('Anonymous User', self.stream.getvalue())
 
         user.assertCalled()
@@ -35,7 +30,7 @@ class Test(CLITestCase):
         content = json.dumps({'login': 'eggs', 'created_at': '1/2/2000'})
         user = urls.register(method='GET', path='/user', content=content)
 
-        main(['--show-traceback', 'whoami'], False)
+        main(['--show-traceback', 'whoami'])
         self.assertIn('eggs', self.stream.getvalue())
 
         user.assertCalled()
@@ -48,16 +43,12 @@ class Test(CLITestCase):
         os.environ.pop('ANACONDA_API_TOKEN', None)
 
         # requests.get_netrc_auth uses expanduser to find the netrc file, point to our test file
-        expanduser = mock.Mock(return_value=data_dir('netrc'))
-        with mock.patch('os.path.expanduser', expanduser):
+        expanduser = unittest.mock.Mock(return_value=data_dir('netrc'))
+        with unittest.mock.patch('os.path.expanduser', expanduser):
             auth = requests.utils.get_netrc_auth('http://localhost', raise_errors=True)
         self.assertEqual(auth, ('anonymous', 'pass'))
 
         user = urls.register(path='/user', status=401)
 
-        main(['--show-traceback', 'whoami'], False)
+        main(['--show-traceback', 'whoami'])
         self.assertNotIn('Authorization', user.req.headers)
-
-
-if __name__ == '__main__':
-    unittest.main()